Delay sending notifications while it's night
No one wants to be notified while they're sleeping. And when looking at
the notifications when wakeing up, the study count is more accurate if
it's more recent.

Using the timezone information to calculate nightime and delay.

Users don't get notified from 9pm to 7am for now.

+++ brain/profile.go
@@ -18,13 +18,13 @@ type profile struct {
 type profileData struct {
 	Name      string
 	Locale    string
-	Timezone  float64
+	Timezone  int
 	CacheTime time.Time
 }
 
-func (p profile) Name() string      { return p.data.Name }
-func (p profile) Locale() string    { return p.data.Locale }
-func (p profile) Timezone() float64 { return p.data.Timezone }
+func (p profile) Name() string   { return p.data.Name }
+func (p profile) Locale() string { return p.data.Locale }
+func (p profile) Timezone() int  { return p.data.Timezone }
 
 // GetProfile fetches a cached profile.
 // Returns ErrNotFound if none found or cache is older than profileMaxCacheTime.
