+++ common/profile.go
@@ -6,5 +6,5 @@ package common
 type Profile interface {
 	Name() string
 	Locale() string
-	Timezone() float64
+	Timezone() int
 }
